+++ src/js/types/html.js
@@ -7,7 +7,7 @@ class HtmlType {
     this.userSettings = userSettings
 
     const TARGET_SELECTOR = el.hasAttribute('data-target') ? el.getAttribute('data-target') : el.getAttribute('href')
-    const TARGET = document.querySelector(TARGET_SELECTOR).cloneNode(true)
+    const TARGET = document.querySelector(TARGET_SELECTOR)
 
     if (!TARGET) {
       throw new Error(`Ups, I can't find the target ${TARGET_SELECTOR}.`)
@@ -34,18 +34,14 @@ class HtmlType {
         VIDEO.currentTime = VIDEO.getAttribute('data-time')
       }
 
-      if (this.userSettings.autoplayVideo) {
-        // Start playback (and loading if necessary)
-        VIDEO.play()
-      }
+      // Start playback (and loading if necessary)
+      VIDEO.play()
     }
 
     const audio = container.querySelector('audio')
     if (audio) {
-      if (this.userSettings.autoplayAudio) {
-        // Start playback (and loading if necessary)
-        audio.play()
-      }
+      // Start playback (and loading if necessary)
+      audio.play()
     }
 
     container.classList.add('tobii-group-' + group)